+++ src/DetailsView/reports/components/report-sections/report-body.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 
 import { NamedSFC } from '../../../../common/react/named-sfc';
-import { CheckListProps, DetailsProps, ReportSectionFactory, SummaryProps } from './automated-checks-report-section-factory';
+import { CheckListProps, DetailsProps, ReportSectionFactory, SummaryProps } from './report-section-factory';
 
 export type ReportBodyProps = {
     sectionFactory: ReportSectionFactory;
