+++ src/DetailsView/reports/components/report-sections/report-section-factory.tsx
